requirements-prod.txt
# Production Dependencies (Minimal)
# Install with: pip install -r requirements-prod.txt

# Web Framework & API
streamlit==1.24.0
fastapi==0.100.0
uvicorn==0.22.0
requests==2.31.0

# Data Models & Validation
pydantic==2.0.3
pydantic-settings==2.0.3

# AI & LLM Integration
openai==1.3.0
azure-openai==1.0.0b1
langchain==0.0.350
langchain-openai==0.0.2
langgraph==0.0.20

# Database & Storage
# RDFLib Knowledge Graph (Neo4j 대체)
rdflib==7.0.0
rdflib-sqlalchemy==0.5.0
rdflib-jsonld==0.6.1
redis==5.0.1
sqlalchemy==2.0.23

# Vector Store & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.4

# Search & Data Collection
google-search-results==2.4.2
duckduckgo-search==3.9.6
beautifulsoup4==4.12.2

# UI Components & Visualization  
streamlit-agraph==0.0.45
plotly==5.17.0

# Template & Document Processing
jinja2==3.1.2
markdown==3.5.1

# Authentication & Security
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.6

# Communication & Notifications
slack-sdk==3.26.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
aiohttp==3.9.1
tenacity==8.2.3
structlog==23.2.0
pandas==2.1.4
jsonschema==4.20.0